backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.0
chromadb>=0.4.18
onnxruntime>=1.16.0
requests>=2.31.0
pypdf>=3.17.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic>=2.0.0
googlesearch-python>=1.2.3
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
